search bar with results and debouncer
