+++ package.json
@@ -19,6 +19,9 @@
   "contributors": [
     {
       "name": "Mariana Meireles"
+    },
+    {
+      "name": "Madhur Tandon"
     }
   ],
   "files": [
@@ -51,6 +54,7 @@
     "watch:labextension": "jupyter labextension watch ."
   },
   "dependencies": {
+    "@jupyter-notebook/react-components": "^0.6.2",
     "@jupyterlab/application": "^3.1.0",
     "@jupyterlab/apputils": "^3.1.9",
     "@jupyterlab/coreutils": "^5.1.0",
